Rename config functions and files to utils, update imports and config loading.
