+++ src/lib/config/utils/loadConfig.ts
@@ -1,11 +1,12 @@
-import { loadEnvConfig } from './services/env'
-import { loadGitConfig } from './services/git'
-import { loadGitignore, loadIgnore } from './services/ignore'
-import { loadProjectJsonConfig } from './services/project'
-import { loadXDGConfig } from './services/xdg'
-import { Config } from './types'
+import { loadEnvConfig } from '../services/env'
+import { loadGitConfig } from '../services/git'
+import { loadGitignore, loadIgnore } from '../services/ignore'
+import { loadProjectJsonConfig } from '../services/project'
+import { loadXDGConfig } from '../services/xdg'
+import { Config } from '../types'
 
-import { DEFAULT_CONFIG } from './constants'
+import { DEFAULT_CONFIG } from '../constants'
+import { BaseCommandOptions } from '../../../commands/types'
 
 /**
  * Load application config
@@ -24,15 +25,18 @@ import { DEFAULT_CONFIG } from './constants'
  *
  * @returns {Config} application config
  **/
-export function loadConfig(argv = {}): Config {
+export function loadConfig<ConfigType, ArgvType = BaseCommandOptions>(
+  argv = {} as ArgvType
+) {
   // Default config
   let config = DEFAULT_CONFIG
+
   config = loadGitignore(config)
   config = loadIgnore(config)
   config = loadXDGConfig(config)
   config = loadGitConfig(config)
   config = loadProjectJsonConfig(config)
   config = loadEnvConfig(config)
-  
-  return { ...config, ...argv }
+
+  return { ...config, ...argv } as Config & ConfigType & ArgvType
 }
